0.3 : Improve master password, Bug overflow on big backup, Add Tools : Create Check Unlock

+++ resticterm/Models/Backup.cs
@@ -18,10 +18,10 @@ namespace resticterm.Models
         {
             public string message_type { get; set; }
             public float percent_done { get; set; }
-            public int total_files { get; set; }
-            public int files_done { get; set; }
-            public int total_bytes { get; set; }
-            public int bytes_done { get; set; }
+            public long total_files { get; set; }
+            public long files_done { get; set; }
+            public long total_bytes { get; set; }
+            public long bytes_done { get; set; }
         }
 
         /// <summary>
@@ -30,17 +30,17 @@ namespace resticterm.Models
         public class Summary
         {
             public string message_type { get; set; }
-            public int files_new { get; set; }
-            public int files_changed { get; set; }
-            public int files_unmodified { get; set; }
-            public int dirs_new { get; set; }
-            public int dirs_changed { get; set; }
-            public int dirs_unmodified { get; set; }
-            public int data_blobs { get; set; }
-            public int tree_blobs { get; set; }
-            public int data_added { get; set; }
-            public int total_files_processed { get; set; }
-            public int total_bytes_processed { get; set; }
+            public long files_new { get; set; }
+            public long files_changed { get; set; }
+            public long files_unmodified { get; set; }
+            public long dirs_new { get; set; }
+            public long dirs_changed { get; set; }
+            public long dirs_unmodified { get; set; }
+            public long data_blobs { get; set; }
+            public long tree_blobs { get; set; }
+            public long data_added { get; set; }
+            public long total_files_processed { get; set; }
+            public long total_bytes_processed { get; set; }
             public float total_duration { get; set; }
             public string snapshot_id { get; set; }
         }
